Update Overview Page and ViewModel with new layout

+++ B4.EE.KarlstromB/B4.EE.KarlstromB/ViewModels/OverviewViewModel.cs
@@ -16,14 +16,19 @@ namespace B4.EE.KarlstromB.ViewModels
     {
         private readonly IAppSettingsService settingsService;
         private readonly ICocktailsService cocktailsService;
-
+        private ISeederService seederService;
 
         public OverviewViewModel(
-            IAppSettingsService settingsService, 
-            ICocktailsService cocktailsService)
+            IAppSettingsService settingsService,
+            ICocktailsService cocktailsService,
+            ISeederService seederService)
         {
             this.settingsService = settingsService;
             this.cocktailsService = cocktailsService;
+            this.seederService = seederService;
+
+            this.seederService.EnsureSeeded();
+            Cocktails = new ObservableCollection<Cocktail>(cocktailsService.GetAllCocktails().Result);
         }
 
         private ObservableCollection<Cocktail> cocktails;
@@ -40,25 +45,34 @@ namespace B4.EE.KarlstromB.ViewModels
         }
 
         public ICommand OpenCocktailPageCommand => new Command<Cocktail>(
-            async (Cocktail cocktail) => {
+            async (Cocktail cocktail) =>
+            {
                 await CoreMethods.PushPageModel<CocktailViewModel>(cocktail, false, true);
             }
         );
 
         public ICommand DeleteCocktailCommand => new Command<Cocktail>(
-            async (Cocktail cocktail) => {
-                await cocktailsService.DeleteCocktail(cocktail.Id);
-                await RefreshCocktails();
+            async (Cocktail cocktail) =>
+            {
+                bool answer = await CoreMethods.DisplayAlert($"Delete {cocktail.Name}", $"Are you sure you want to delete {cocktail.Name}?", "Ok", "Cancel");
+
+                if (answer)
+                {
+                    await cocktailsService.DeleteCocktail(cocktail.Id);
+                    Cocktails.Remove(cocktail);
+                    await RefreshCocktails();
+                }
+                else
+                {
+                    return;
+                }
             }
         );
 
         private async Task RefreshCocktails()
         {
-            var settings = await settingsService.GetSettings();
-            var cocktails = await cocktailsService.GetCocktailsForUser(settings.CurrentUserId);
-            Cocktails = null;
+            var cocktails = await cocktailsService.GetAllCocktails();
             Cocktails = new ObservableCollection<Cocktail>(cocktails.OrderBy(e => e.Name));
         }
-        //public IList<Cocktail> CocktailExamples { get { return BasicCocktails.CocktailsExamples; } }
     }
 }
